First refactoring; XmlConfigParser class created; Control class refactored; ReportReader class adjusted (recently created); AboutWindow updated; AUTHORS dist file updated; ChangeLog dist file updated;

+++ RS/src/Model/Analysis/ReportReader.cpp
@@ -1,3 +1,36 @@
+/* -*- Mode: C; indent-tabs-mode: t; c-basic-offset: 4; tab-width: 4 -*-  */
+/*
+* ReportReader.cpp
+* Copyright (C) 2014-2016 LEDS - Univali <zeferino@univali.br>
+* Laboratory of Embedded and Distributed Systems
+* University of Vale do Itajaí
+*
+* Authors: Laboratory of Embedded and Distributed Systems (LEDS)
+*
+* Contact: Profº Dr. Cesar Albenes Zeferino {zeferino@univali.br}
+*
+* RedScarf is free software: you can redistribute it and/or modify it
+* under the terms of the GNU General Public License as published by the
+* Free Software Foundation, either version 3 of the License, or
+* (at your option) any later version.
+*
+* RedScarf is distributed in the hope that it will be useful, but
+* WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+* See the GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License along
+* with this program.  If not, see <http://www.gnu.org/licenses/>.
+*
+* ------------------------------ Reviews -------------------------------------
+* Date       - Version - Author                      | Description
+* ----------------------------------------------------------------------------
+* 31/05/2016 - 1.1     - Eduardo Alves da Silva      | First refactoring
+*    ||      - ||      - Sérgio Vargas Junior        |      ||
+* ----------------------------------------------------------------------------
+*
+*/
+
 #include "include/Model/Analysis/ReportReader.h"
 #include "include/Model/Analysis/DataReport.h"
 
@@ -48,61 +81,61 @@ std::list<DataReport *>* ReportReader::readReport(const char *file) {
             scanResult = fscanf(fpIn,"%s", str);
             if(scanResult != EOF) {
 
-                fClk = atof(str);
+                fClk = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
                 accNbOfPck = (unsigned long int) atol(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                avgRequiredBwNorm = atof(str);
+                avgRequiredBwNorm = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                acceptedTrafficFlits = atof(str);
+                acceptedTrafficFlits = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                idealAvgLatency = atof(str);
+                idealAvgLatency = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                avgLatencyCycles = atof(str);
+                avgLatencyCycles = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                stdevLatency = atof(str);
+                stdevLatency = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                minLatency = atof(str);
+                minLatency = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                maxLatency = atof(str);
+                maxLatency = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                avgRequiredBwBps = atof(str);
+                avgRequiredBwBps = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                acceptedTrafficBps = atof(str);
+                acceptedTrafficBps = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                idealAvgLatencyNs = atof(str);
+                idealAvgLatencyNs = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                avgLatencyNs = atof(str);
+                avgLatencyNs = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                minLatencyNs = atof(str);
+                minLatencyNs = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                maxLatencyNs = atof(str);
+                maxLatencyNs = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                metDeadlinesPer0 = atof(str);
+                metDeadlinesPer0 = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                metDeadlinesPer1 = atof(str);
+                metDeadlinesPer1 = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                metDeadlinesPer2 = atof(str);
+                metDeadlinesPer2 = (float) atof(str);
 
                 scanResult = fscanf(fpIn,"%s", str);
-                metDeadlinesPer3 = atof(str);
+                metDeadlinesPer3 = (float) atof(str);
 
 
                 data->fClk                  = fClk;
